[refactor] : api 상수 파일을 constants 폴더에서 service 폴더로 이동
